requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
djangorestframework
markdown
djangorestframework-simplejwt==3.2.3
django-filter
factory_boy
django-extensions
jupyter
googlemaps
geopy
polyline